pyproject.toml
[tool.poetry]
name = "pageloot"
version = "0.1.0"
description = ""
authors = ["Sipan <sipanm19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10,<3.12"
python-dotenv = "^1.0.0"
cachecontrol = "^0.12.7"
gunicorn = "20.1.0"
certifi = "^2022.12.7"
requests = "^2.29.0"
python-multipart = "^0.0.6"
psycopg2-binary = "^2.9.6"
django = "^5.1.1"
djangorestframework = "^3.15.2"
drf-yasg = "^1.21.7"
django-filter = "^24.3"
factory-boy = "^3.3.1"
pytest-django = "^4.9.0"


[tool.poetry.dev-dependencies]
black = "^23.3.0"
pytest-mock = "3.10.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
